requirements.txt
django >= 3
django-mptt >= 0.11
django-ckeditor >= 5.9
beautifulsoup4
requests